revert factoring out cli
reverting as we are not using a build.rs script anymore
to generate shell completion with clap. This is now done
in the print-x-setup script in someone's rc file
